backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
executing==2.0.1
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.12.1
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-core==2.4.1
google-cloud-storage==2.14.0rc1
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.62.0
idna==3.6
imageio==2.33.1
install==1.3.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
kiwisolver==1.4.5
kombu==5.3.4
MarkupSafe==2.1.3
# matplotlib==3.8.2
# matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
packaging==23.2
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.1.0a1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
torch
torchaudio
torchvision
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12